requirements.txt
django==5.2.6
requests
pandas
numpy
pandas_ta
python-binance
pybit
python-decouple
